+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Releases prior to January 2023 are tracked on the project GitHub [Releases Page]
 
 ### Changed
 
+### Fixed
+
+## 1.403 - 2026-05-06
+
+### Changed
+
 - Arbitrary JARs in `lib/modules` do not automatically become part of the `classpath` in `IdePlugin`. Any such JAR must belong to a plugin content module that is explicitly declared in the `plugin.xml` ([#1479](https://github.com/JetBrains/intellij-plugin-verifier/pull/1479))  
 - Detect duplicate entries in ZIPs and JARs
 - Reduce plugin-structure allocation and FS overhead in plugin loading paths ([#1476](https://github.com/JetBrains/intellij-plugin-verifier/pull/1476))
@@ -19,7 +25,7 @@ Releases prior to January 2023 are tracked on the project GitHub [Releases Page]
 ### Fixed
 
 - Fix NPE during ProductModuleV2 construction when missing classPath field ([#1470](https://github.com/JetBrains/intellij-plugin-verifier/pull/1470), [MP-8060](https://youtrack.jetbrains.com/issue/MP-8060))
-- Fix path traversal in ZIP decompression on Windows filesystems 
+- Fix path traversal in ZIP decompression on Windows filesystems
 
 ## 1.402 - 2026-04-01